# openapi.json
{"openapi":"3.1.0","info":{"title":"Aviator Signals API","version":"0.1.0"},"paths":{"/admin":{"get":{"summary":"Admin Page","operationId":"admin_page_admin_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/me":{"get":{"summary":"Admin Me","operationId":"admin_me_admin_api_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/login":{"post":{"summary":"Admin Login","operationId":"admin_login_admin_api_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/api/logout":{"post":{"summary":"Admin Logout","operationId":"admin_logout_admin_api_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/metrics":{"get":{"summary":"Admin Metrics","operationId":"admin_metrics_admin_api_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/logs":{"get":{"summary":"Admin Logs","operationId":"admin_logs_admin_api_logs_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":80,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/api/logs/clear":{"post":{"summary":"Admin Logs Clear","operationId":"admin_logs_clear_admin_api_logs_clear_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/notify":{"post":{"summary":"Admin Notify","operationId":"admin_notify_admin_api_notify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminNotify"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/api/subs":{"get":{"summary":"Admin List Subs","operationId":"admin_list_subs_admin_api_subs_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/stats":{"get":{"summary":"Get Today Stats","operationId":"get_today_stats_api_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/ultimo-historico":{"get":{"summary":"Api Ultimo Historico","operationId":"api_ultimo_historico_api_ultimo_historico_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/ultimo-resultado":{"get":{"summary":"Api Ultimo Resultado","operationId":"api_ultimo_resultado_api_ultimo_resultado_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/velas":{"get":{"summary":"Api Velas Get","operationId":"api_velas_get_api_velas_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Api Velas Post","operationId":"api_velas_post_api_velas_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/health":{"get":{"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/sinal":{"post":{"summary":"Post Sinal","operationId":"post_sinal_api_sinal_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Sinal"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/resultado":{"post":{"summary":"Post Resultado","operationId":"post_resultado_api_resultado_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Resultado"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/notification/click":{"post":{"summary":"Notification Click","description":"Registra clique em notificação push","operationId":"notification_click_api_notification_click_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/subscribe":{"post":{"summary":"Subscribe","operationId":"subscribe_api_subscribe_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/test/push-resultado":{"post":{"summary":"Test Push Resultado","operationId":"test_push_resultado_api_test_push_resultado_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/push-broadcast":{"post":{"summary":"Push Broadcast","operationId":"push_broadcast_api_push_broadcast_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/subs/reset":{"post":{"summary":"Subs Reset","operationId":"subs_reset_api_subs_reset_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/online":{"get":{"summary":"Api Online","operationId":"api_online_api_online_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/stream":{"get":{"summary":"Stream","operationId":"stream_api_stream_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/analytics/overview":{"get":{"summary":"Admin Analytics Overview","description":"Retorna visão geral de analytics","operationId":"admin_analytics_overview_admin_api_analytics_overview_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/analytics/visits":{"get":{"summary":"Admin Analytics Visits","description":"Retorna acessos por dia (últimos 7 dias)","operationId":"admin_analytics_visits_admin_api_analytics_visits_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/analytics/devices":{"get":{"summary":"Admin Analytics Devices","description":"Retorna distribuição de dispositivos","operationId":"admin_analytics_devices_admin_api_analytics_devices_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/campaigns":{"get":{"summary":"Admin Campaigns","description":"Retorna histórico de campanhas manuais","operationId":"admin_campaigns_admin_api_campaigns_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/notifications/stats":{"get":{"summary":"Admin Notifications Stats","description":"Retorna estatísticas gerais de notificações","operationId":"admin_notifications_stats_admin_api_notifications_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/analytics/advanced":{"get":{"summary":"Admin Analytics Advanced","description":"Retorna analytics avançados:\n- Tendências (↑↓%)\n- Retenção e tempo médio\n- Novos vs recorrentes\n- Alertas automáticos","operationId":"admin_analytics_advanced_admin_api_analytics_advanced_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/analytics/hourly":{"get":{"summary":"Admin Analytics Hourly","description":"Retorna distribuição de visitas por hora (para heatmap)","operationId":"admin_analytics_hourly_admin_api_analytics_hourly_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/api/analytics/visits/detailed":{"get":{"summary":"Admin Analytics Visits Detailed","description":"Retorna analytics detalhados de visitas com filtros de período.\nperiod: 7d, 30d, 90d, max, custom\nstart_date/end_date: formato YYYY-MM-DD (para period=custom)","operationId":"admin_analytics_visits_detailed_admin_api_analytics_visits_detailed_get","parameters":[{"name":"period","in":"query","required":false,"schema":{"type":"string","default":"7d","title":"Period"}},{"name":"start_date","in":"query","required":false,"schema":{"type":"string","title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"type":"string","title":"End Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AdminLogin":{"properties":{"user":{"type":"string","title":"User"},"password":{"type":"string","title":"Password"}},"type":"object","required":["user","password"],"title":"AdminLogin"},"AdminNotify":{"properties":{"title":{"type":"string","maxLength":120,"minLength":1,"title":"Title"},"body":{"type":"string","maxLength":280,"minLength":1,"title":"Body"},"kind":{"type":"string","title":"Kind","description":"admin|sinal|resultado|aviso","default":"admin"},"tag":{"anyOf":[{"type":"string","maxLength":120},{"type":"null"}],"title":"Tag"},"url":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"title":"Url","default":"/"},"priority":{"type":"integer","maximum":10.0,"minimum":0.0,"title":"Priority","default":5},"mode":{"type":"string","title":"Mode","description":"queue|sync","default":"queue"},"target":{"type":"string","title":"Target","description":"all|sample|contains|host","default":"all"},"limit":{"type":"integer","maximum":500.0,"minimum":0.0,"title":"Limit","default":0},"query":{"anyOf":[{"type":"string","maxLength":120},{"type":"null"}],"title":"Query"},"delay_seconds":{"type":"integer","maximum":86400.0,"minimum":0.0,"title":"Delay Seconds","default":0},"dry_run":{"type":"boolean","title":"Dry Run","default":false},"renotify":{"type":"boolean","title":"Renotify","default":true},"require_interaction":{"type":"boolean","title":"Require Interaction","default":false},"silent":{"type":"boolean","title":"Silent","default":false}},"type":"object","required":["title","body"],"title":"AdminNotify"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Resultado":{"properties":{"id":{"type":"string","title":"Id"},"status":{"type":"string","title":"Status"},"vela_final":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Vela Final"},"ts":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ts"}},"type":"object","required":["id","status"],"title":"Resultado"},"Sinal":{"properties":{"tipo":{"type":"string","title":"Tipo","default":"entrada_confirmada"},"apos_de":{"type":"number","title":"Apos De"},"cashout":{"type":"number","title":"Cashout"},"max_gales":{"type":"integer","title":"Max Gales","default":2},"vela_atual":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Vela Atual"},"meta":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Meta"},"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"ts":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ts"}},"type":"object","required":["apos_de","cashout"],"title":"Sinal"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}